requirements.txt
praw==7.7.1
boto3==1.34.14
python-dotenv==1.0.0
pmaw==3.0.0
pandas==2.2.3
apache-airflow-providers-amazon==10.0.0
apache-airflow==2.11.0
